requirements.txt
# Librerías para desarrollo local
pylint==2.15.2
autopep8==1.7.0
pre-commit==2.20.0

# Librerías para producción
crispy-bootstrap5==0.6
django==4.0.7
django-colorfield==0.7.2
django-guardian==2.4.0
django-environ==0.9.0
djangorestframework==3.13.1
coreapi

django-filter==22.1
django-extensions==3.2.0
markdown==3.4.1
django-dotenv
django-storages[azure]

# Libreria para exportal la información

reportlab
openpyxl
pandas
# GDAL==3.3.1

# # Librerías para conectar con postgres
# pg_config
psycopg2-binary


# Librerías para internacionalización y localización
django-rosetta==0.9.8
django-modeltranslation

# Librerías de conexión con SQL Server
pyodbc
mssql-django==1.1.3
azure-mgmt-sql
python-decouple

# Pendientes por revisar implementación
# django-phonenumber-field==7.0.0

# Librería para imágenes
django-any-imagefield==1.1

# Librería de países
pycountry==22.3.5

# Libreria para conectar con react
django-cors-headers
redis
django-filter == 22.1.0

# librerias para el dashboard map
channels==4.0.0
channels-redis==4.2.0
aioredis==2.0.1
daphne==3.0.2

# Añadir gunicorn para el servidor web
gunicorn
twisted[http2,tls]
